split multi action
