+++ crates/client/src/app/input/mod.rs
@@ -2,3 +2,4 @@ pub mod body;
 pub mod crafting;
 pub mod hotbar_assignment;
 pub mod inventory;
+pub mod multi_action;
